11.27

+++ User/hardware.h
@@ -10,8 +10,8 @@
 
 float BMYBJ_YB(u16 init_freq, u8 init_temp, u16 freq, u8 temp);
 float MRYBJ_YB(u16 init_freq, u8 init_temp, u16 freq, u8 temp);
-float GJJ_YL(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t K);
-float TYLH_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t K);
-float SYJ_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, int32_t A, int32_t B, int32_t C, int32_t b);
-float MSJ_YL_KN(u16* init_freq, u8 init_temp, u16* freq, u8 temp, u8 size, int32_t a, int32_t b);
+float GJJ_YL(u16 init_freq, u8 init_temp, u16 freq, u8 temp, double K);
+float TYLH_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, double K);
+float SYJ_YL_MPa(u16 init_freq, u8 init_temp, u16 freq, u8 temp, double A, double B, double C, double b);
+float MSJ_YL_KN(u16* init_freq, u8 init_temp, u16* freq, u8 temp, u8 size, double a, double b);
 #endif //ISD2190_HARDWARE_H
